// src/app/users/models/users.ts
export class Users {

  userId: string = '';
  username: string = '';
  userFirstname: string = '';
  userLastname: string = '';
  password: string = '';
  createdDate : string = '';
  previousLogin: string = '';
  status: string = '';
  userGender: string = '';
  userAddress: string = '';
  email: string = '';
  userPhone: string = '';
  userTeacherRef: string = ''
  userType: string = '';
  userSupervisor: string = '';
}